refactor(cart): :recycle: movido componente de Carrinho para pasta de components
